.fname_regex {shinyAce}	R Documentation

Regular expression for matching the function name in a completion line in the middle of a function call

Description

Regular expression for matching the function name in a completion line in the middle of a function call

Usage

.fname_regex

Format

An object of class character of length 1.

[Package shinyAce version 0.4.4 Index]