re_capture {shinyAce}	R Documentation

Retrieve regular expression named capture groups as a list

Description

Retrieve regular expression named capture groups as a list

Usage

re_capture(x, re, ...)

Arguments

x
a character string to capture from

re
the regular expression to use

...
additional arguments passed to regexpr

Value

a named list of matches

Examples

shinyAce:::re_capture("ak09j b", "(?<num>\\d+)(?<alpha>[a-zA-Z]+)", perl = TRUE)

[Package shinyAce version 0.4.4 Index]