Update Ace Editor

Update the styling or mode of an aceEditor component.

updateAceEditor(session, editorId, value, theme, readOnly, mode, fontSize,
  wordWrap, useSoftTabs, tabSize, showInvisibles, border = c("normal",
  "alert", "flash"), autoComplete = c("disabled", "enabled", "live"),
  autoCompleters = c("snippet", "text", "keyword", "static", "rlang"),
  autoCompleteList = NULL)

Arguments

session
The Shiny session to whom the editor belongs

editorId
The ID associated with this element

value
The initial text to be contained in the editor.

theme
The Ace theme to be used by the editor. The theme in Ace determines the styling and coloring of the editor. Use getAceThemes to enumerate all the themes available.

readOnly
If set to TRUE, Ace will disable client-side editing. If FALSE (the default), it will enable editing.

mode
The Ace mode to be used by the editor. The mode in Ace is often the programming or markup language that you're using and determines things like syntax highlighting and code folding. Use the getAceModes function to enumerate all the modes available.

fontSize
If set, will update the font size (in px) used in the editor. Should be an integer.

wordWrap
If set to TRUE, Ace will enable word wrapping. Default value is FALSE.

useSoftTabs
Replace tabs by spaces. Default value is TRUE

tabSize
Set tab size. Default value is 4

showInvisibles
Show invisible characters (e.g., spaces, tabs, newline characters). Default value is FALSE

border
Set the border 'normal', 'alert', or 'flash'.

autoComplete
Enable/Disable code completion. See aceEditor for details.

autoCompleters
List of completers to enable. If set to NULL, all completers will be disabled.

autoCompleteList
If set to NULL, exisitng static completions list will be unset. See aceEditor for details.

Examples

# NOT RUN {
 shinyServer(function(input, output, session) {
   observe({
     updateAceEditor(session, "myEditor", "Updated text for editor here",
       mode = "r", theme = "ambiance")
   })
 }
# }